% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_responses.R
\name{chat_completion_to_tibble}
\alias{chat_completion_to_tibble}
\title{Convert chat completion response to a tibble}
\usage{
chat_completion_to_tibble(response, other_data = NULL)
}
\arguments{
\item{response}{A list containing a chat completion response from the OpenAI API.}

\item{other_data}{other metadata to include in the resulting tibble, for example temperature etc.}
}
\value{
A tibble containing the relevant information from the chat completion response.
}
\description{
This function takes a single chat completion response from the OpenAI API
and converts it into a tibble.
}
\examples{
\dontrun{
  # Assuming `result` is a single response from the OpenAI API
  tibble_result <- chat_completion_to_tibble(result)
}
}
