% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_responses.R
\name{chat_completions_to_tibble}
\alias{chat_completions_to_tibble}
\title{Convert multiple chat completions to a tibble}
\usage{
chat_completions_to_tibble(completions, other_data = NULL)
}
\arguments{
\item{completions}{A list containing multiple chat completion responses from the OpenAI API.}

\item{other_data}{other metadata to include in the resulting tibble, for example temperature etc.}
}
\value{
A tibble containing the chat completion responses.
}
\description{
This function takes a list of chat completion responses from the OpenAI API
and converts them into a single tibble.
}
