% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_model.R
\name{default_model}
\alias{default_model}
\title{Default Model Selection}
\usage{
default_model()
}
\value{
A character string representing the default model identifier.
}
\description{
This function returns the default model identifier for use in certain applications.
The default model is set to "gpt-3.5-turbo-1106".
}
\examples{
default_model()

}
