% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_response.R
\name{get_text_response}
\alias{get_text_response}
\title{Get plain text OpenAI API Response with Retry Logic}
\usage{
get_text_response(
  api_key,
  prompt_list,
  model = default_model(),
  temp = 0,
  max_tries = 5
)
}
\arguments{
\item{api_key}{A character string representing the OpenAI API key for authentication.}

\item{prompt_list}{A list of messages forming the conversation prompt, where each message is a list with "role" (system, user, assistant) and "content" (message text).}

\item{model}{(Optional) A character string specifying the GPT-3 model to use. Defaults to the default model.}

\item{temp}{(Optional) A numeric value representing the temperature for sampling in the range (0, 1]. Defaults to 0.}

\item{max_tries}{(Optional) An integer specifying the maximum number of retry attempts. Defaults to 5.}
}
\value{
A parsed response if successful, otherwise NA.
}
\description{
This function attempts to obtain a response from the OpenAI GPT-3 API with retry logic.
}
\examples{
\dontrun{
# Example usage:
api_key <- "your_openai_api_key"
prompt_list <- list(list(role = "system", content = "You are a drop bear"),
                    list(role = "human", content = "What do you think of koalas?"))
response <- get_text_response(api_key, prompt_list)
}

}
\seealso{
\code{\link{request_chat}}, \code{\link{parse_response}}, \code{\link{increment_attempt}}
}
\keyword{api}
\keyword{logic}
\keyword{openai}
\keyword{response}
\keyword{retry}
