% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json_response.R
\name{parse_json_response}
\alias{parse_json_response}
\title{Parse OpenAI API JSON Response}
\usage{
parse_json_response(resp)
}
\arguments{
\item{resp}{A response object obtained from the OpenAI API call.}
}
\value{
Parsed content from the response if successful, otherwise NA.
}
\description{
This function parses the response received from the OpenAI GPT-3 API.
}
\seealso{
\code{\link{strip_json_tags}}, \code{\link[httr]{content}}
}
\keyword{api}
\keyword{openai}
\keyword{parsing}
\keyword{response}
