% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Parse OpenAI API Response}
\usage{
parse_response(resp)
}
\arguments{
\item{resp}{A response object obtained from the OpenAI API call.}
}
\value{
Parsed content from the response if successful, otherwise NA.
}
\description{
This function parses the response received from the OpenAI GPT-3 API.
}
\examples{
\dontrun{
# Example usage:
api_key <- "your_openai_api_key"
prompt_list <- list(list(role = "system", content = "you are a drop bear"),
                    list(role = "human", content = "What do you think of koalas?"))
response <- request_chat(api_key, prompt_list)
parsed_content <- parse_response(response)
}

}
\seealso{
\code{\link[httr]{content}}
}
\keyword{api}
\keyword{openai}
\keyword{parsing}
\keyword{response}
